requirements.txt
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
Django==2.0
django-extensions==1.9.8
HTMLParser==0.0.2
idna==2.6
pkg-resources==0.0.0
py-sdvb==2.3
PyStarDict==0.7
python-mpd2==0.5.5
pytz==2017.3
requests==2.18.4
six==1.11.0
typing==3.6.2
urllib3==1.22

pytest-django==3.1.2
pytest-sugar==0.8.0
pytest-cov
model_mommy

django-debug-toolbar